Rename module plugin_loader -> loaders
